+++ BackEnd/src/main/java/com/manhattan/busyness_predictor/service/HistoryService.java
@@ -1,14 +1,15 @@
 package com.manhattan.busyness_predictor.service;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import com.manhattan.busyness_predictor.model.History;
 import com.manhattan.busyness_predictor.model.Location;
 import com.manhattan.busyness_predictor.repository.HistoryRepository;
 import com.manhattan.busyness_predictor.repository.LocationRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class HistoryService {
@@ -19,7 +20,7 @@ public class HistoryService {
     @Autowired
     private LocationRepository locationRepository;
 
-    public void addToHistory(Long userId, Long locationId) {
+    public void addToHistory(Integer userId, Integer locationId) {
         // Check if location exists
         locationRepository.findById(locationId)
                 .orElseThrow(() -> new RuntimeException("Location not found"));
@@ -28,7 +29,7 @@ public class HistoryService {
         historyRepository.save(history);
     }
 
-    public List<Location> getSearchHistory(Long userId) {
+    public List<Location> getSearchHistory(Integer userId) {
         List<History> history = historyRepository.findByUserIdOrderByTimestampDesc(userId);
 
         return history.stream()
@@ -39,7 +40,7 @@ public class HistoryService {
                 .collect(Collectors.toList());
     }
 
-    public void clearHistory(Long userId) {
+    public void clearHistory(Integer userId) {
         List<History> userHistory = historyRepository.findByUserId(userId);
         historyRepository.deleteAll(userHistory);
     }
